issue #121 fixes:
added few "LGTM ignores" to prevent alerts when "correcting" them will hurt the correctness of the class or method.
added casting in Statistic's "sampleVariance" function to prevent overflow caused by int multiplication.
added casting in (other)Statistic's "of" function to prevent overflow caused by int multiplication.

improvments:
Some Javadocs caused an error in LGTM because they used "@param < T >" instead of "@param <T>",fixed it.
Some Javadocs caused an error in LGTM because they used the wrong param name (for example, v instead of ¢),fixed it.
ignored implicit conversion from Array to String (because it is created by the Spartanizer)

+++ src/main/java/il/org/spartan/tables/Statistic.java
@@ -81,6 +81,6 @@ public enum Statistic {
   /** @param p percents of quartile (for q1 - 25, q3 - 75, median - 50) */
   static double quartile(final int p, final double[] ds) {
     Arrays.sort(ds);
-    return ds[Math.round(p * ds.length / 100)];
+    return ds[Math.round((float)p * (float)ds.length / 100)];
   }
 }
